- Moved Login and Signup to components directory
